+++ src/sequence/constant.rs
@@ -2,30 +2,30 @@ use crate::sequence::models::Sequence;
 use std::any::type_name;
 
 // Implementirajte konstantno zaporedje
-pub struct Constant <i64> {
-    vrednost : i64,
+pub struct Constant <f64> {
+    vrednost : f64,
 
 }
 
-impl Sequence<i64> for Constant<i64> {
+impl Sequence<f64> for Constant<f64> {
     fn name(&self) -> String {
         format!("Const: {}",self.vrednost)
     }
-    fn contains(&self, item: i64) -> bool {
+    fn contains(&self, item: f64) -> bool {
         if item == self.vrednost {
             true
         }
         else {false}
     }
-    fn k_th(&self, k: usize) -> Option<i64> {
+    fn k_th(&self, k: usize) -> Option<f64> {
         Some(self.vrednost)
     }
-    fn start(&self) -> i64 {
+    fn start(&self) -> f64 {
         self.vrednost
     }
 }
-impl Constant<i64> {
-    pub fn new(x: i64) -> Constant<i64> {
+impl Constant<f64> {
+    pub fn new(x: f64) -> Constant<f64> {
         Constant {vrednost: x}
     }
 }
